p2papp/requirements.txt
pycrypto
pyjwt